inventree-bambulab-farm/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "inventree-bambulab-farm"
version = "0.1.0"
description = "InvenTree plugin for monitoring a Bambu Lab printer farm"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
  { name = "3dprinterv2 contributors" }
]
dependencies = [
  "paho-mqtt>=2.1.0",
]

[project.urls]
Homepage = "https://github.com/example/inventree-bambulab-farm"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
